+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.1.5 [unreleased]
 
+- Added duplicate indexes
 - Added maintenance tab
 - Added `table_caching` and `index_caching` methods
 - Added configurable cache hit rate threshold